+++ Source/Get-ChocolateyPackagePath.ps1
@@ -2,6 +2,12 @@
 .SYNOPSIS
 Gets the path where the Chocolatey package is currently being installed or uninstalled from.
 
+.DESCRIPTION
+
+Chocolatey has provided a 'chocolateyPackageFolder' environment variable since
+v0.9.8.20 (late 2012), so you should use that instead, unless you have a need to
+support really old chocolatey clients (if that's even possible).
+
 .EXAMPLE
 
 PS> $packagePath = Get-ChocolateyPackagePath
@@ -14,6 +20,11 @@ Gets the package folder that the current chocolatey script is being run from.
 param(
     [string]$CommandPath=$MyInvocation.PSCommandPath
 )
+
+if ($env:chocolateyPackageFolder) {
+    Write-Host "HINT: Use the 'chocolateyPackageFolder' environment variable instead." -ForegroundColor Magenta
+    return $env:chocolateyPackageFolder
+}
 
 $commandName = Split-Path $CommandPath -Leaf